+++ TICKET/core/src/main/java/com/toy/core/domain/performance/PerformanceScheduleRepository.java
@@ -7,4 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface PerformanceScheduleRepository extends JpaRepository<PerformanceSchedule, Long> {
 
     List<PerformanceSchedule> findByPerformanceId(Long performanceId);
+
+    List<PerformanceSchedule> findByPerformanceIdIn(List<Long> performanceIds);
 }